requirements.txt
crewai>=0.11.0
streamlit>=1.29.0
python-dotenv>=1.0.0
langchain>=0.1.0
beautifulsoup4>=4.12.2
markdown>=3.5.1
pyyaml>=6.0.1
requests>=2.31.0
python-frontmatter>=1.0.0
langchain-community==0.3.14
google-api-python-client>=2.100.0
python-frontmatter>=1.1.0
langchain-google-community==2.0.4